Add build command + move insecure to base. Fixes #62
